# API/API.yaml
openapi: 3.0.0
info:
  title: Athena MHMD API
  description: >-
    An API for multi-level privacy preservation techniques.
  version: 0.0.1
...
paths:
...
  /smpc/histogram/:
    post:
      tags:
        - SMPC
      summary: Initiate a secure histogram computation with desired parameters.
      description: "Compute histogram(s) on the specified attributes, based on specified datasources. A filter is applied before the histogram(s) computation. Returns the requested histogram(s) in JSON format."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Histogram"
      responses:
        '202':
          description: Histogram computation accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComputationAccepted"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /smpc/queue:
    get:
      tags:
        - SMPC
      summary: Poll for an ongoing computation status.
      parameters:
        - in: query
          name: request
          schema:
            type: integer
          description: The request id.
      responses:
        '200':
          description: ΟΚ.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComputationResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  schemas:
...
    Error:
      properties:
        code:
          type: integer
          description: Error code
        message:
          type: string
          description: Error message
...
    Histogram:
      required:
        - attributes
      properties:
        attributes:
          description: "Attributes for the histogram(s) to be computed"
          type: "array"
          items:
            type: "array"
            items:
              type: "object"
              properties:
                name:
                  type: "string"
                  description: "Attribute's name."
                cells:
                  type: "integer"
                  description: "Number of equi-width cells for that attribute."
          example: [ [ {"name":"Age", "cells":5} ],[ {"name": "Height", cells: 3}, {"name": "Weight", cells: 4}, {"name": "Heart Rate", cells: 3}] ]
        datasources:
          description: "Datasources from which the histogram(s) will be computated. Optional. All datasources will be considered if left blank."
          type: "array"
          items:
            type: "string"
          example: ["HospitalA", "HospitalB"]
        filters:
          description: "List of conditions that should be met in order to take a record into consideration in a secure histogram computation. All conditions are joined with the same specified operator. Optional."
          type: "object"
          properties:
            operator:
              type: "string"
              description: "Boolean Binary operator performed on the conditions"
              enum:
              - "AND"
              - "OR"
              - "XOR"
            conditions:
              type: "array"
              items:
                type: "object"
                properties:
                  attribute:
                    type: "string"
                  operator:
                    type: "string"
                    enum:
                    - ">"
                    - "<"
                    - "="
                  value:
                    type: "string"
              example: [{"attribute":"Sex", "operator":"=", "value":"Female" }, {"attribute":"Age", "operator":">", "value":"18" }]
    ComputationResponse:
      required:
        - status
      description: "List of serialized versions for each histogram."
      type: object
      properties:
        status:
          type: string
          enum:
            - "succeeded"
            - "running"
            - "failed"
            - "notstarted"
        step:
          description: "Current step of the computation"
          type: string
        result:
          type: array
          items:
            type: object
            description: "Serialized versions for the computed histogram."
            properties:
              cellsPerDimension:
                type: array
                items:
                  type: integer
              histogram:
                type: array
                items:
                  type: integer
          example: [{ "cellsPerDimension": [5], "histogram": [237, 211, 239, 161, 152] }, {"cellsPerDimension": [3, 4, 3], "histogram": [17, 2, 0, 17, 22, 3, 2, 55, 50, 0, 34, 24, 48, 14, 0, 66, 51, 11, 5, 142, 97, 0, 102, 98, 9, 3, 0, 16, 14, 2, 1, 30, 32, 0, 11, 22] } ]
    ComputationAccepted:
      required:
        - location
      description: "Indication that the secure histogram computation has been accepted. The location provided should be used for status polling, and for results retrieval."
      type: object
      properties:
        location:
          type: string
      example: {"location" : "/smpc/queue?request=1"}